+++ app/routes/app.tsx
@@ -52,7 +52,6 @@ export default function App() {
         </s-link>
         <s-link href="/app/rules">{t.rules}</s-link>
         <s-link href="/app/messages">{t.messages}</s-link>
-        <s-link href="/app/plan">{t.plan}</s-link>
         <s-link href="/app/guide">{t.guide}</s-link>
       </s-app-nav>
       <Outlet />